cleanup: move code into modules and simplify
